Make OIDCLoginInitView exempt from xframe protection
I am integrating django-lti into a django application, and connecting it
to Canvas LMS. From my testing, this view needs to be exempt from
XFrameOptionsMiddleware if this middleware is turned on.

The application author can easily get around this by sub-classing
OIDCLoginInitView themselves and making this alteration in the
application, but I don't see a reason why this shouldn't be in
django-lti already.

This is related to ticket #126.